refactor: split DiffPane into diff/ subcomponents
Pure structural extraction — behavior, props, data-testids, and class
strings unchanged. DiffPane.tsx is now a ~185-line orchestrator; hunk
rendering, highlighting, selection derivation, and toolbar pieces live
in src/content/overlay/components/diff/.
